Final version

+++ laba3.c
@@ -3,20 +3,20 @@
 #include<math.h>
 #include<time.h>
 
-int fib_recur(int N){
+unsigned long long	fib_recur(int N){
 	if(N == 1 || N == 2){
 		return 1;
 	}
-	return fib_recur(N - 1) + fib_recur( N -2);
+	return fib_recur(N - 1) + fib_recur(N - 2);
 }
 
-long int fib_formula(int N){
+unsigned long long	fib_formula(int N){
 	long double fib;
 	fib = (long int)(powl((1 + sqrtl(5)) / 2, (long double)N) / sqrtl(5));
 	return fib; 
 }
 
-long int fib_mass(int N){
+unsigned long long	fib_mass(int N){
 	long int mass[1000];
 	mass[0] = 0;
 	mass[1] = 1;
@@ -26,7 +26,7 @@ long int fib_mass(int N){
 	return mass[N];
 }
 
-long int fib_matrix(int N){
+unsigned long long	fib_matrix(int N){
 	long int mass[2][2] = {{1, 1}, {1, 0}}, mass1[2][2] = {{1, 1}, {1, 0}}, mass2[2][2] = {{1, 1}, {1, 0}}, rez;
 	for(int i = 1; i < N; i++){
 		mass[0][0] = mass1[0][0] * mass2[0][0] + mass1[0][1] * mass2[1][0];
@@ -50,25 +50,25 @@ int main(){
 	scanf("%d", &N);
 
 	start = clock();
-	printf("Рекурсия : %d : ", fib_recur(N));
+	printf("Рекурсия : %llu : ", fib_recur(N));
 	end = clock();
 	seconds = (double)(end - start) / CLOCKS_PER_SEC;
 	printf("%f seconds \n", seconds);
 
 	start = clock();
-	printf("Формула  : %ld : ", fib_formula(N));
+	printf("Формула  : %llu : ", fib_formula(N));
 	end = clock();
 	seconds = (double)(end - start) / CLOCKS_PER_SEC;
 	printf("%f seconds \n", seconds);
 
 	start = clock();
-	printf("Массив   : %ld : ", fib_mass(N));
+	printf("Массив   : %llu : ", fib_mass(N));
 	end = clock();
 	seconds = (double)(end - start) / CLOCKS_PER_SEC;
 	printf("%f seconds \n", seconds);
 
 	start = clock();
-	printf("Матрицы  : %ld : ", fib_matrix(N));
+	printf("Матрицы  : %llu : ", fib_matrix(N));
 	end = clock();
 	seconds = (double)(end - start) / CLOCKS_PER_SEC;
 	printf("%f seconds \n", seconds);
